backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-filter==24.3
psycopg2-binary==2.9.10
gunicorn==23.0.0
celery==5.4.0
redis==5.2.1
django-redis==5.4.0
dj-database-url==2.3.0
openai==1.58.1
anthropic==0.40.0
langchain==0.3.13
langchain-openai==0.3.0
langchain-anthropic==0.3.3
langchain-core==0.3.28
python-dotenv==1.0.1
whitenoise==6.8.2
django-extensions==3.2.3
Markdown==3.7
pyyaml==6.0.2
